requirements.txt
streamlit
ultralytics
opencv-python-headless
numpy
pillow
streamlit-image-comparison